requirements.txt
apache-airflow[pandas]
boto3
minio
apache-airflow-providers-mysql
